+++ packages/astro/src/content/vite-plugin-content-virtual-mod.ts
@@ -24,9 +24,6 @@ export function astroContentVirtualModPlugin({
 	);
 	const contentEntryExts = getContentEntryExts(settings);
 
-	const assetsDir = settings.config.experimental.assets
-		? contentPaths.assetsDir.toString()
-		: 'undefined';
 	const extGlob =
 		contentEntryExts.length === 1
 			? // Wrapping {...} breaks when there is only one extension
@@ -38,14 +35,8 @@ export function astroContentVirtualModPlugin({
 		.replace('@@CONTENT_DIR@@', relContentDir)
 		.replace('@@ENTRY_GLOB_PATH@@', entryGlob)
 		.replace('@@RENDER_ENTRY_GLOB_PATH@@', entryGlob);
-	const virtualAssetsModContents = fsMod
-		.readFileSync(contentPaths.virtualAssetsModTemplate, 'utf-8')
-		.replace('@@ASSETS_DIR@@', assetsDir);
 
 	const astroContentVirtualModuleId = '\0' + VIRTUAL_MODULE_ID;
-	const allContents = settings.config.experimental.assets
-		? virtualModContents + virtualAssetsModContents
-		: virtualModContents;
 
 	return {
 		name: 'astro-content-virtual-mod-plugin',
@@ -58,7 +49,7 @@ export function astroContentVirtualModPlugin({
 		load(id) {
 			if (id === astroContentVirtualModuleId) {
 				return {
-					code: allContents,
+					code: virtualModContents,
 				};
 			}
 		},