+++ packages/astro/src/core/dev/container.ts
@@ -71,7 +71,7 @@ export async function createContainer(params: CreateContainerParams = {}): Promi
 		logging,
 		isRestart,
 	});
-	const { host, headers } = settings.config.server;
+	const { host, headers, open } = settings.config.server;
 
 	// The client entrypoint for renderers. Since these are imported dynamically
 	// we need to tell Vite to preoptimize them.
@@ -82,14 +82,9 @@ export async function createContainer(params: CreateContainerParams = {}): Promi
 	const viteConfig = await createVite(
 		{
 			mode: 'development',
-			server: { host, headers },
+			server: { host, headers, open },
 			optimizeDeps: {
 				include: rendererClientEntries,
-			},
-			define: {
-				'import.meta.env.BASE_URL': settings.config.base
-					? JSON.stringify(settings.config.base)
-					: 'undefined',
 			},
 		},
 		{ settings, logging, mode: 'dev', command: 'dev', fs }