+++ packages/astro/test/core-image.test.js
@@ -97,13 +97,19 @@ describe('astro:image', () => {
 				expect(res.status).to.equal(200);
 			});
 
+			it('returns proper content-type', async () => {
+				let $img = $('#local img');
+				let src = $img.attr('src');
+				let res = await fixture.fetch(src);
+				expect(res.headers.get('content-type')).to.equal('image/webp');
+			});
+
 			it('errors on unsupported formats', async () => {
 				logs.length = 0;
 				let res = await fixture.fetch('/unsupported-format');
 				await res.text();
 
 				expect(logs).to.have.a.lengthOf(1);
-				console.log(logs[0].message);
 				expect(logs[0].message).to.contain('Received unsupported format');
 			});
 		});
@@ -212,6 +218,19 @@ describe('astro:image', () => {
 				let $img = $('img');
 				expect($img.attr('src').startsWith('/_image')).to.equal(true);
 			});
+
+			it('properly handles remote images', async () => {
+				let res = await fixture.fetch('/httpImage');
+				let html = await res.text();
+				$ = cheerio.load(html);
+
+				let $img = $('img');
+				expect($img).to.have.a.lengthOf(2);
+				const remoteUrls = ['https://example.com/image.png', '/image.png'];
+				$img.each((index, element) => {
+					expect(element.attribs['src']).to.equal(remoteUrls[index]);
+				});
+			});
 		});
 
 		describe('getImage', () => {
@@ -244,12 +263,25 @@ describe('astro:image', () => {
 
 			it('Adds the <img> tags', () => {
 				let $img = $('img');
-				expect($img).to.have.a.lengthOf(4);
+				expect($img).to.have.a.lengthOf(7);
 			});
 
 			it('has proper source for directly used image', () => {
 				let $img = $('#direct-image img');
 				expect($img.attr('src').startsWith('/src/')).to.equal(true);
+			});
+
+			it('has proper source for refined image', () => {
+				let $img = $('#refined-image img');
+				expect($img.attr('src').startsWith('/src/')).to.equal(true);
+			});
+
+			it('has proper sources for array of images', () => {
+				let $img = $('#array-of-images img');
+				const imgsSrcs = [];
+				$img.each((i, img) => imgsSrcs.push(img.attribs['src']));
+				expect($img).to.have.a.lengthOf(2);
+				expect(imgsSrcs.every((img) => img.startsWith('/src/'))).to.be.true;
 			});
 
 			it('has proper attributes for optimized image through getImage', () => {
@@ -289,6 +321,129 @@ describe('astro:image', () => {
 			it('includes /src in the path', async () => {
 				expect($('img').attr('src').startsWith('/src')).to.equal(true);
 			});
+		});
+	});
+
+	describe('proper errors', () => {
+		/** @type {import('./test-utils').DevServer} */
+		let devServer;
+		/** @type {Array<{ type: any, level: 'error', message: string; }>} */
+		let logs = [];
+
+		before(async () => {
+			fixture = await loadFixture({
+				root: './fixtures/core-image-errors/',
+				experimental: {
+					assets: true,
+				},
+			});
+
+			devServer = await fixture.startDevServer({
+				logging: {
+					level: 'error',
+					dest: new Writable({
+						objectMode: true,
+						write(event, _, callback) {
+							logs.push(event);
+							callback();
+						},
+					}),
+				},
+			});
+		});
+
+		after(async () => {
+			await devServer.stop();
+		});
+
+		it("properly error when getImage's first parameter isn't filled", async () => {
+			logs.length = 0;
+			let res = await fixture.fetch('/get-image-empty');
+			await res.text();
+
+			expect(logs).to.have.a.lengthOf(1);
+			expect(logs[0].message).to.contain('Expected getImage() parameter');
+		});
+
+		// TODO: For some reason, this error crashes the dev server?
+		it.skip('properly error when src is undefined', async () => {
+			logs.length = 0;
+			let res = await fixture.fetch('/get-image-undefined');
+			await res.text();
+
+			expect(logs).to.have.a.lengthOf(1);
+			expect(logs[0].message).to.contain('Expected src to be an image.');
+		});
+
+		it('properly error image in Markdown frontmatter is not found', async () => {
+			logs.length = 0;
+			let res = await fixture.fetch('/blog/one');
+			await res.text();
+
+			expect(logs).to.have.a.lengthOf(1);
+			expect(logs[0].message).to.contain('does not exist. Is the path correct?');
+		});
+
+		it('properly error image in Markdown content is not found', async () => {
+			logs.length = 0;
+			let res = await fixture.fetch('/post');
+			await res.text();
+
+			expect(logs).to.have.a.lengthOf(1);
+			expect(logs[0].message).to.contain('Could not find requested image');
+		});
+	});
+
+	describe('support base option correctly', () => {
+		before(async () => {
+			fixture = await loadFixture({
+				root: './fixtures/core-image-base/',
+				experimental: {
+					assets: true,
+				},
+				base: '/blog',
+			});
+			await fixture.build();
+		});
+
+		it('has base path prefix when using the Image component', async () => {
+			const html = await fixture.readFile('/index.html');
+			const $ = cheerio.load(html);
+			const src = $('#local img').attr('src');
+			expect(src.length).to.be.greaterThan(0);
+			expect(src.startsWith('/blog')).to.be.true;
+		});
+
+		it('has base path prefix when using getImage', async () => {
+			const html = await fixture.readFile('/get-image/index.html');
+			const $ = cheerio.load(html);
+			const src = $('img').attr('src');
+			expect(src.length).to.be.greaterThan(0);
+			expect(src.startsWith('/blog')).to.be.true;
+		});
+
+		it('has base path prefix when using image directly', async () => {
+			const html = await fixture.readFile('/direct/index.html');
+			const $ = cheerio.load(html);
+			const src = $('img').attr('src');
+			expect(src.length).to.be.greaterThan(0);
+			expect(src.startsWith('/blog')).to.be.true;
+		});
+
+		it('has base path prefix in Markdown', async () => {
+			const html = await fixture.readFile('/post/index.html');
+			const $ = cheerio.load(html);
+			const src = $('img').attr('src');
+			expect(src.length).to.be.greaterThan(0);
+			expect(src.startsWith('/blog')).to.be.true;
+		});
+
+		it('has base path prefix in Content Collection frontmatter', async () => {
+			const html = await fixture.readFile('/blog/one/index.html');
+			const $ = cheerio.load(html);
+			const src = $('img').attr('src');
+			expect(src.length).to.be.greaterThan(0);
+			expect(src.startsWith('/blog')).to.be.true;
 		});
 	});