+++ packages/astro/test/fixtures/dynamic-endpoint-collision/src/pages/api/catch/[...slug].ts
@@ -1,0 +1,15 @@
+import type { APIRoute } from "astro";
+
+const slugs = ["one", undefined];
+
+export const get: APIRoute = ({ params }) => {
+  return {
+    body: JSON.stringify({
+      slug: params.slug || "index",
+    }),
+  };
+};
+
+export function getStaticPaths() {
+  return slugs.map((u) => ({ params: { slug: u } }));
+}